Finished SPotifySdkService and update of RoomManager and TrackManager
