+++ music_room_app/lib/spotify_library/track/managers/track_image_manager.dart
@@ -2,17 +2,14 @@ import 'package:flutter/cupertino.dart';
 import 'package:music_room_app/home/models/track.dart';
 import 'package:music_room_app/services/spotify_service_subscriber.dart';
 import 'package:spotify_sdk/models/player_state.dart';
-import 'package:spotify_sdk/models/track.dart';
-import '../../../services/spotify_sdk_static.dart';
+import '../../../services/spotify_sdk_service.dart';
 
 class TrackImageManager with ChangeNotifier implements TrackManager {
-  TrackImageManager({required this.trackApp, required this.tracksList});
+  TrackImageManager({required this.trackApp, required this.tracksList, required this.spotify});
 
   final List<TrackApp> tracksList;
+  SpotifySdkService spotify;
   TrackApp trackApp;
-  Track? trackSdk;
-  String? get trackSdkId =>
-      trackSdk == null ? null : trackSdk!.uri.split(':')[2];
 
   Widget? returnImage() {
     if (trackApp.album != null) {
@@ -29,13 +26,9 @@ class TrackImageManager with ChangeNotifier implements TrackManager {
 
   @override
   void whenPlayerStateChange(PlayerState newState) {
-    trackSdk = newState.track;
-    if (trackSdk != null) {
-      String? newId = trackSdkId;
-      if (trackApp.id != newId) {
-        trackApp = SpotifySdkStatic.findNewTrackApp(trackApp, tracksList, newId);
+      if (trackApp.id != spotify.currentTrack?.id) {
+        trackApp = spotify.currentTrack ?? trackApp;
         notifyListeners();
       }
     }
-  }
 }
