+++ music_room_app/lib/spotify_library/track/managers/track_slider_row_manager.dart
@@ -3,7 +3,8 @@ import 'package:flutter/cupertino.dart';
 import 'package:music_room_app/services/spotify_service_subscriber.dart';
 import 'package:spotify_sdk/models/player_state.dart';
 import 'package:spotify_sdk/models/track.dart';
-import '../../../services/spotify_sdk_static.dart';
+
+import '../../../services/spotify_sdk_service.dart';
 
 class TrackSliderRowManager with ChangeNotifier implements TrackManager {
   TrackSliderRowManager();
